# lib/api-zod/src/generated/types/downloadTokenRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadTokenRequest {
  /** Optional legacy field — access is granted when the CV is marked paid server-side. */
  paymentId?: string;
}