# lib/api-zod/src/generated/types/paystackInitializeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CV Builder API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaystackInitializeInput {
  email: string;
}